requirements.txt
dynaconf==3.1.12
fastapi==0.99.0
PyGithub==1.59.*
retry==0.9.2
Jinja2==3.1.2
tiktoken==0.4.0
uvicorn==0.22.0
pytest==7.4.0
aiohttp==3.8.4
atlassian-python-api==3.39.0
GitPython==3.1.32
PyYAML==6.0
starlette-context==0.3.6
boto3==1.28.25
google-cloud-storage==2.10.0
ujson==5.8.0
azure-devops==7.1.0b3
msrest==0.7.1
##
openai
litellm
duckdb
datasets
notebook
black
evaluate
click
code_contests_tester==0.1.6
aiolimiter
Jinja2
tqdm
pysnooper
loguru
numpy
retry

# uninstall ipython to catch breakpoints on debug with sandbox==false